requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
datasets>=2.14.0
duckduckgo-search>=6.0.0
ddgs>=9.0.0
python-dotenv>=1.0.0
huggingface-hub>=0.20.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
streamlit>=1.30.0
